# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dunda club operating system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products for the POS
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
...
  /staff:
    get:
      operationId: getStaff
      tags: [staff]
      summary: List staff members
      responses:
        "200":
          description: Staff members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffInput"
      responses:
        "201":
          description: Created staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
...
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: List events
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: List inventory items
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: adjustInventory
      tags: [inventory]
      summary: Adjust inventory quantity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustInventoryInput"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Get sales report
      parameters:
        - in: query
          name: dateFrom
          schema:
            type: string
            format: date
        - in: query
          name: dateTo
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
  /reports/products:
    get:
      operationId: getProductReport
      tags: [reports]
      summary: Get product sales report
      parameters:
        - in: query
          name: dateFrom
          schema:
            type: string
            format: date
        - in: query
          name: dateTo
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Product report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductReportItem"
  /audit-logs:
    get:
      operationId: getAuditLogs
      tags: [audit]
      summary: List audit log entries
      parameters:
        - in: query
          name: entity
          schema:
            type: string
        - in: query
          name: action
          schema:
            type: string
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
...
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: List payments
      parameters:
        - in: query
          name: orderId
          schema: { type: string }
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Payment" }
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment against an order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentInput" }
      responses:
        "201":
          description: Recorded payment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
...
  /reports/inventory:
    get:
      operationId: getInventoryReport
      tags: [reports]
      summary: Inventory position, movements and discrepancies
      parameters:
        - in: query
          name: branchId
          schema: { type: string }
      responses:
        "200":
          description: Inventory report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryReport" }
  /reports/staff:
    get:
      operationId: getStaffReport
      tags: [reports]
      summary: Staff sales and shift performance
      parameters:
        - in: query
          name: branchId
          schema: { type: string }
      responses:
        "200":
          description: Staff report
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StaffReportItem" }
  /reports/tables:
    get:
      operationId: getTableReport
      tags: [reports]
      summary: Table utilisation and revenue
      parameters:
        - in: query
          name: branchId
          schema: { type: string }
      responses:
        "200":
          description: Table report
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TableReportItem" }
  /reports/payments:
    get:
      operationId: getPaymentReport
      tags: [reports]
      summary: Payment method breakdown and refunds
      parameters:
        - in: query
          name: branchId
          schema: { type: string }
      responses:
        "200":
          description: Payment report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentReport" }
  /reports/events:
    get:
      operationId: getEventReport
      tags: [reports]
      summary: Event performance
      parameters:
        - in: query
          name: branchId
          schema: { type: string }
      responses:
        "200":
          description: Event report
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EventReportItem" }
...
components:
...
  schemas:
...
    Product:
      type: object
      required:
        [
          id,
          name,
          categoryId,
          category,
          price,
          unit,
          baseUnit,
          stock,
          available,
          active,
          trackInventory,
          accent,
        ]
      properties:
        id: { type: string }
        name: { type: string }
        categoryId: { type: string }
        category: { type: string }
        price: { type: number }
        unit: { type: string }
        baseUnit: { type: string }
        stock: { type: number }
        available: { type: boolean }
        active: { type: boolean }
        trackInventory: { type: boolean }
        accent: { type: string }
        sku: { type: ["string", "null"] }
        barcode: { type: ["string", "null"] }
        cost: { type: integer }
        tax: { type: integer }
        description: { type: ["string", "null"] }
...
    StaffMember:
      type: object
      required: [id, name, email, role, status]
      properties:
        id: { type: string }
        clerkUserId: { type: string }
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        role: { type: string }
        roleId: { type: ["string", "null"] }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        branchId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    CreateStaffInput:
      type: object
      required: [name, email, roleId]
      properties:
        clerkUserId: { type: ["string", "null"] }
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        roleId: { type: string }
        branchId: { type: ["string", "null"] }
...
    Event:
      type: object
      required: [id, name, date, startTime, endTime, capacity, status]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        date: { type: string, format: date }
        startTime: { type: string }
        endTime: { type: string }
        capacity: { type: integer }
        status: { type: string, enum: [DRAFT, UPCOMING, LIVE, COMPLETED, CANCELLED] }
        revenue: { type: number }
    CreateEventInput:
      type: object
      required: [name, date, startTime, endTime, capacity]
      properties:
        name: { type: string }
        date: { type: string, format: date }
        startTime: { type: string }
        endTime: { type: string }
        capacity: { type: integer, minimum: 1 }
        description: { type: ["string", "null"] }
        branchId: { type: ["string", "null"] }
...
    InventoryItem:
      type: object
      required: [id, name, currentQuantity, reorderLevel, unit]
      properties:
        id: { type: string }
        productId: { type: ["string", "null"] }
        name: { type: string }
        category: { type: ["string", "null"] }
        sku: { type: ["string", "null"] }
        currentQuantity: { type: number }
        reorderLevel: { type: number }
        cost: { type: ["number", "null"] }
        unit: { type: string }
    AdjustInventoryInput:
      type: object
      required: [productId, quantity, type]
      properties:
        productId: { type: string }
        quantity: { type: number }
        type: { type: string, enum: [PURCHASE, SALE, ADJUSTMENT, WASTE, RETURN, STOCK_COUNT] }
        reason: { type: ["string", "null"] }
...
    AuditLog:
      type: object
      required: [id, action, entity, createdAt]
      properties:
        id: { type: string }
        staffId: { type: ["string", "null"] }
        staffName: { type: ["string", "null"] }
        action: { type: string, enum: [LOGIN, CREATE, UPDATE, DELETE, VOID, REFUND, ADJUST, APPROVE, DISCARD] }
        entity: { type: string }
        entityId: { type: ["string", "null"] }
        detail: { type: ["string", "null"] }
        reason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...
    SalesReport:
      type: object
      required: [dateFrom, dateTo, totalRevenue, totalOrders, averageOrderValue, byPaymentMethod, byHour]
      properties:
        dateFrom: { type: string, format: date }
        dateTo: { type: string, format: date }
        totalRevenue: { type: number }
        totalOrders: { type: integer }
        averageOrderValue: { type: number }
        byPaymentMethod:
          type: array
          items:
            $ref: "#/components/schemas/MetricPoint"
        byHour:
          type: array
          items:
            $ref: "#/components/schemas/MetricPoint"
    ProductReportItem:
      type: object
      required: [name, category, quantitySold, revenue]
      properties:
        name: { type: string }
        category: { type: string }
        quantitySold: { type: integer }
        revenue: { type: number }
    MetricPoint:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: number }
...
    ProductInput:
      type: object
      required: [name, categoryId, category, price]
      properties:
        name: { type: string, minLength: 1 }
        categoryId: { type: string }
        category: { type: string }
        baseUnit: { type: string }
        sku: { type: ["string", "null"] }
        barcode: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        cost: { type: integer, minimum: 0 }
        price: { type: integer, minimum: 0 }
        tax: { type: integer, minimum: 0 }
        trackInventory: { type: boolean }
        accent: { type: string }
...
    Payment:
      type: object
      required: [id, organizationId, branchId, orderId, amount, method, status, paidAt]
      properties:
        id: { type: string }
        organizationId: { type: string }
        branchId: { type: string }
        orderId: { type: string }
        amount: { type: integer }
        method: { type: string }
        reference: { type: ["string", "null"] }
        status: { type: string }
        cashierId: { type: ["string", "null"] }
        paidAt: { type: string, format: date-time }
        outstanding: { type: integer }
        fullyPaid: { type: boolean }
    PaymentInput:
      type: object
      required: [orderId, amount, method]
      properties:
        orderId: { type: string }
        amount: { type: integer, minimum: 1 }
        method:
          type: string
          enum: [CASH, MPESA, CARD, BANK_TRANSFER, OTHER]
        reference: { type: ["string", "null"] }
        isFinal: { type: boolean }
...
    InventoryReport:
      type: object
      required:
        [totalItems, stockValue, belowReorder, outOfStock, byMovement, discrepancies]
      properties:
        totalItems: { type: integer }
        stockValue: { type: number }
        belowReorder: { type: integer }
        outOfStock: { type: integer }
        byMovement:
          type: array
          items: { $ref: "#/components/schemas/MetricPoint" }
        discrepancies:
          type: array
          items:
            type: object
            properties:
              productId: { type: string }
              name: { type: ["string", "null"] }
              variance: { type: number }
              unit: { type: ["string", "null"] }
    StaffReportItem:
      type: object
      required: [staffId, name, orders, revenue, averageOrderValue]
      properties:
        staffId: { type: ["string", "null"] }
        name: { type: string }
        orders: { type: integer }
        revenue: { type: number }
        averageOrderValue: { type: number }
    TableReportItem:
      type: object
      required: [id, name, section, seats, status, revenue, covers, averageStayMinutes]
      properties:
        id: { type: string }
        name: { type: string }
        section: { type: string }
        seats: { type: integer }
        status: { type: string }
        revenue: { type: number }
        covers: { type: integer }
        averageStayMinutes: { type: integer }
    PaymentReport:
      type: object
      required: [total, refunded, byMethod]
      properties:
        total: { type: number }
        refunded: { type: number }
        byMethod:
          type: array
          items:
            type: object
            required: [label, value, count]
            properties:
              label: { type: string }
              value: { type: number }
              count: { type: integer }
...